01-faq-bot/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
langgraph==0.2.60
langchain-anthropic==0.3.1
python-dotenv==1.0.1
